refactor: group components by domain
ticket/, list/, filters/, settings/ and shared/ mirror the desk's
domain folders. Vue filenames are untouched — a studio component's name
IS its filename, and every block JSON references it by that name — so
only import paths move. The filter subsystem gets a README and its
helper modules drop the redundant kbFilter prefix now that the folder
carries the context.
