+++ studio/helpdesk/components/settings/KbOrganizationAvatars.vue
